+++ fordead/steps/step3_dieback_detection.py
@@ -181,7 +181,8 @@ def dieback_detection(
             write_tif(stress_data["nb_dates"], first_detection_date_index.attrs,tile.paths["nb_dates_stress"],nodata=0)
             del stress_data
 
-        # print("Détection du déperissement")
+    # update tile info with new anomalies    
+    tile.getdict_datepaths("Anomalies",tile.paths["AnomaliesDir"]) # Get paths and dates to previously calculated anomalies
     tile.save_info()