+++ fordead/validation_process.py
@@ -230,9 +230,9 @@ def filter_cloudy_acquisitions(reflect, cloudiness_path, lim_perc_cloud):
     return reflect
     
 def source_mask_dataframe(source_mask, sentinel_source):
-    if sentinel_source=="THEIA":
+    if sentinel_source in ["theia", "theiastac"]:
         source_mask = source_mask>0
-    elif sentinel_source=="Scihub" or sentinel_source=="PEPS":
+    elif sentinel_source in ["scihub", "peps", "Planetary"]:
         source_mask = ~source_mask.isin([4,5])
     return source_mask